SM2 Algorithm Update: Each grade now results distinct relative interval
Original SM-2 only affects the efactor of the next recommendation and does nothing to the interval. This is strange because if something was harder I'd expect to review it again soon and visa versa. This follows the philosophy of Anki's tweaks to SM2 and gives users full control of both efactor and interval respectively.