Bug Fix: Missing attributes when switching review mode
This fixes an issue where practice attributes are missing when switching from fixed interval card to spaced card.

The issue is that we're getting the latest practice data from the practiceData field. This works in most cases since we generate a new card for all cards without any sessions. However in this case, we already have a fixed session so latest session was incorrect.

To fix I refactored this so that we pass in the latest session info to practice handler ensuring that this is never wrong.